rename SDK folders
