+++ examples/user/hyperdbg_app/header/pch.h
@@ -32,4 +32,4 @@
 // HyperDbg SDK headers
 //
 #include "SDK/HyperDbgSdk.h"
-#include "SDK/Imports/user1/HyperDbgLibImports.h"
+#include "SDK/imports1/user/HyperDbgLibImports.h"
