+++ hyperdbg/hyperlog/header/pch.h
@@ -40,8 +40,8 @@
 
 #include "UnloadDll.h"
 #include "SDK/HyperDbgSdk.h"
-#include "SDK/Modules/HyperLog.h"
-#include "SDK/Imports/kernel1/HyperDbgHyperLogImports.h"
+#include "SDK/modules1/HyperLog.h"
+#include "SDK/imports1/kernel/HyperDbgHyperLogImports.h"
 #include "components/spinlock/header/Spinlock.h"
 #include "Logging.h"
