+++ hyperdbg/include/SDK/headers1/BasicTypes.h
