+++ hyperdbg/include/SDK/headers1/Connection.h
