+++ hyperdbg/include/SDK/headers1/ErrorCodes.h
