+++ hyperdbg/include/SDK/headers1/HardwareDebugger.h
