+++ hyperdbg/include/SDK/headers1/RequestStructures.h
