+++ hyperdbg/include/SDK/headers1/ScriptEngineCommonDefinitions.h
