+++ hyperdbg/include/SDK/headers1/Symbols.h
