+++ hyperdbg/include/SDK/imports1/kernel/HyperDbgHyperLogImports.h
