+++ hyperdbg/include/SDK/imports1/kernel/HyperDbgHyperLogIntrinsics.h
