+++ hyperdbg/include/SDK/imports1/kernel/HyperDbgVmmImports.h
