+++ hyperdbg/include/SDK/imports1/user/HyperDbgLibImports.h
