+++ hyperdbg/include/SDK/imports1/user/HyperDbgScriptImports.h
