+++ hyperdbg/include/SDK/imports1/user/HyperDbgSymImports.h
