+++ hyperdbg/include/SDK/modules1/HyperLog.h
